Deprecate all multi commands (#941)
