+++ docs/templates/jfrog-pipelines/pipelines-dotnet.yml
@@ -187,8 +187,8 @@ pipelines:
               getFrogbotScriptPath=$( [[ -z "$JF_RELEASES_REPO" ]] && echo "https://releases.jfrog.io" || echo "${JF_URL}/artifactory/${JF_RELEASES_REPO}" )
               curl -fLg "$getFrogbotScriptPath/artifactory/frogbot/v2/[RELEASE]/getFrogbot.sh" | sh
               restore_cache_files dotnet_cache ~/.nuget/packages
-              ./frogbot scan-all-pull-requests
-              ./frogbot scan-multiple-repositories
+              ./frogbot scan-pull-request
+              ./frogbot scan-repository
               add_cache_files dotnet_cache ~/.nuget/packages || true
 
             # For Windows runner:
@@ -202,6 +202,6 @@ pipelines:
             #   $scriptUrl = "$($getFrogbotScriptPath)/artifactory/frogbot/v2/[RELEASE]/getFrogbot.sh"
             #   Invoke-Expression (Invoke-WebRequest -Uri $scriptUrl -UseBasicParsing).Content
             #   restore_cache_files dotnet_cache "%userprofile%\.nuget\packages"
-            #   .\frogbot.exe scan-all-pull-requests
-            #   .\frogbot.exe scan-multiple-repositories
+            #   .\frogbot.exe scan-pull-request
+            #   .\frogbot.exe scan-repository
             #   add_cache_files dotnet_cache "%userprofile%\.nuget\packages" || true
