+++ docs/templates/jfrog-pipelines/pipelines-pip.yml
@@ -191,8 +191,8 @@ pipelines:
             - |
               getFrogbotScriptPath=$( [[ -z "$JF_RELEASES_REPO" ]] && echo "https://releases.jfrog.io" || echo "${JF_URL}/artifactory/${JF_RELEASES_REPO}" )
               curl -fLg "$getFrogbotScriptPath/artifactory/frogbot/v2/[RELEASE]/getFrogbot.sh" | sh
-              ./frogbot scan-all-pull-requests
-              ./frogbot scan-multiple-repositories
+              ./frogbot scan-pull-request
+              ./frogbot scan-repository
 
             # For Windows runner:
             # - |
@@ -204,5 +204,5 @@ pipelines:
             #
             #   $scriptUrl = "$($getFrogbotScriptPath)/artifactory/frogbot/v2/[RELEASE]/getFrogbot.sh"
             #   Invoke-Expression (Invoke-WebRequest -Uri $scriptUrl -UseBasicParsing).Content
-            #   .\frogbot.exe scan-all-pull-requests
-            #   .\frogbot.exe scan-multiple-repositories
+            #   .\frogbot.exe scan-pull-request
+            #   .\frogbot.exe scan-repository
