--- testdata/scanallpullrequests/test-proj-with-vulnerability/package.json
@@ -1,8 +1,0 @@
-{
-  "name": "test-proj",
-  "version": "1.0.0",
-  "dependencies": {
-    "minimatch": "0.2.14",
-    "minimist":"1.2.5"
-  }
-}
