+++ utils/outputwriter/testsutils.go
@@ -13,8 +13,7 @@ import (
 
 var (
 	// Used for tests that are outside the outputwriter package.
-	TestMessagesDir       = filepath.Join("..", "testdata", "messages")
-	TestSummaryCommentDir = filepath.Join(TestMessagesDir, "summarycomment")
+	TestMessagesDir = filepath.Join("..", "testdata", "messages")
 	// Used for tests that are inside the outputwriter package.
 	testMessagesDir       = filepath.Join("..", TestMessagesDir)
 	testReviewCommentDir  = filepath.Join(testMessagesDir, "reviewcomment")
@@ -60,22 +59,4 @@ func GetJsonBodyOutputFromFile(t *testing.T, filePath string) []byte {
 	bytes, err := json.Marshal(bodyRes)
 	assert.NoError(t, err)
 	return bytes
-}
-
-func GetPRSummaryContentNoIssues(t *testing.T, summaryTestDir string, entitled, simplified bool) string {
-	dataPath := filepath.Join(summaryTestDir, "structure")
-	if simplified {
-		if entitled {
-			dataPath = filepath.Join(dataPath, "summary_comment_no_issues_simplified_entitled.md")
-		} else {
-			dataPath = filepath.Join(dataPath, "summary_comment_no_issues_simplified_not_entitled.md")
-		}
-	} else {
-		if entitled {
-			dataPath = filepath.Join(dataPath, "summary_comment_no_issues_pr_entitled.md")
-		} else {
-			dataPath = filepath.Join(dataPath, "summary_comment_no_issues_pr_not_entitled.md")
-		}
-	}
-	return GetOutputFromFile(t, dataPath)
 }
