+++ utils/params.go
@@ -336,7 +336,7 @@ func (g *Git) setDefaultsIfNeeded(gitParamsFromEnv *Git, commandName string) (er
 			return
 		}
 	}
-	if commandName == ScanRepository || commandName == ScanMultipleRepositories {
+	if commandName == ScanRepository {
 		if err = g.extractScanRepositoryEnvParams(gitParamsFromEnv); err != nil {
 			return
 		}
@@ -429,7 +429,7 @@ func GetFrogbotDetails(commandName string) (frogbotDetails *FrogbotDetails, err 
 		return
 	}
 
-	gitParamsFromEnv, err := extractGitParamsFromEnvs(commandName)
+	gitParamsFromEnv, err := extractGitParamsFromEnvs()
 	if err != nil {
 		return
 	}
@@ -489,7 +489,7 @@ func getConfigAggregator(xrayVersion, xscVersion string, gitClient vcsclient.Vcs
 func getConfigFileContent(gitClient vcsclient.VcsClient, gitParamsFromEnv *Git, commandName string) ([]byte, error) {
 	var errMissingConfig *ErrMissingConfig
 
-	if commandName == ScanRepository || commandName == ScanMultipleRepositories {
+	if commandName == ScanRepository {
 		configFileContent, err := ReadConfigFromFileSystem(osFrogbotConfigPath)
 		if err != nil && !errors.As(err, &errMissingConfig) {
 			return nil, err
@@ -572,7 +572,7 @@ func extractJFrogCredentialsFromEnvs() (*coreconfig.ServerDetails, error) {
 	return &server, nil
 }
 
-func extractGitParamsFromEnvs(commandName string) (*Git, error) {
+func extractGitParamsFromEnvs() (*Git, error) {
 	e := &ErrMissingEnv{}
 	var err error
 	gitEnvParams := &Git{}
@@ -608,7 +608,7 @@ func extractGitParamsFromEnvs(commandName string) (*Git, error) {
 	}
 
 	// [Mandatory] Set the repository name, except for multi repository.
-	if err = readParamFromEnv(GitRepoEnv, &gitEnvParams.RepoName); err != nil && commandName != ScanMultipleRepositories {
+	if err = readParamFromEnv(GitRepoEnv, &gitEnvParams.RepoName); err != nil {
 		return nil, err
 	}
